+++ examples/db_remember.rs
@@ -50,9 +50,9 @@ async fn main() {
 
     let handler = Update::filter_message()
         .enter_dialogue::<Message, ErasedStorage<State>, State>()
-        .branch(teloxide::handler![State::Start].endpoint(start))
+        .branch(dptree::case![State::Start].endpoint(start))
         .branch(
-            teloxide::handler![State::GotNumber(n)]
+            dptree::case![State::GotNumber(n)]
                 .branch(dptree::entry().filter_command::<Command>().endpoint(got_number))
                 .branch(dptree::endpoint(invalid_command)),
         );