+++ examples/purchase.rs
@@ -67,19 +67,19 @@ async fn main() {
 fn schema() -> UpdateHandler<Box<dyn std::error::Error + Send + Sync + 'static>> {
     let command_handler = teloxide::filter_command::<Command, _>()
         .branch(
-            teloxide::handler![State::Start]
-                .branch(teloxide::handler![Command::Help].endpoint(help))
-                .branch(teloxide::handler![Command::Start].endpoint(start)),
+            dptree::case![State::Start]
+                .branch(dptree::case![Command::Help].endpoint(help))
+                .branch(dptree::case![Command::Start].endpoint(start)),
         )
-        .branch(teloxide::handler![Command::Cancel].endpoint(cancel));
+        .branch(dptree::case![Command::Cancel].endpoint(cancel));
 
     let message_handler = Update::filter_message()
         .branch(command_handler)
-        .branch(teloxide::handler![State::ReceiveFullName].endpoint(receive_full_name))
+        .branch(dptree::case![State::ReceiveFullName].endpoint(receive_full_name))
         .branch(dptree::endpoint(invalid_state));
 
     let callback_query_handler = Update::filter_callback_query().chain(
-        teloxide::handler![State::ReceiveProductChoice { full_name }]
+        dptree::case![State::ReceiveProductChoice { full_name }]
             .endpoint(receive_product_selection),
     );